Consolidate fake estimator classes/functions in the fake_estimator module
