+++ src/augury/pipelines/pipelines.py
@@ -2,7 +2,7 @@ from typing import Dict
 
 from kedro.pipeline import Pipeline
 
-from tests.fixtures.fake_pipeline import create_fake_pipeline
+from tests.fixtures.fake_estimator import create_fake_pipeline
 from augury.nodes import feature_calculation
 from augury.settings import CATEGORY_COLS
 from .player_pipeline import create_player_pipeline
